Address PR feedback - move reconnect function
