+++ elixir/lib/portal/entra/sync.ex
@@ -59,7 +59,7 @@ defmodule Portal.Entra.Sync do
     delete_unsynced(directory, synced_at)
 
     # Reconnect orphaned policies after sync (groups may have been recreated)
-    reconnected = Portal.Policies.Reconnect.reconnect_orphaned_policies(directory.account_id)
+    reconnected = Portal.Policy.reconnect_orphaned_policies(directory.account_id)
 
     if reconnected > 0 do
       Logger.info("Reconnected #{reconnected} orphaned policies after sync",
